% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_pstrat_data}
\alias{example_pstrat_data}
\title{Create example post-stratification data
#' @description Generates example post-stratification data for MRP analysis.}
\usage{
example_pstrat_data()
}
\value{
An example post-stratification table.
}
\description{
Create example post-stratification data
#' @description Generates example post-stratification data for MRP analysis.
}
