% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_sample_data}
\alias{example_sample_data}
\title{Return example data}
\usage{
example_sample_data(
  is_timevar = TRUE,
  is_aggregated = TRUE,
  special_case = NULL,
  family = "binomial"
)
}
\arguments{
\item{is_timevar}{Logical indicating whether the data is time-varying.}

\item{is_aggregated}{Logical indicating whether the data is aggregated.}

\item{special_case}{Optional character string for specific use cases such as COVID data.
Options are \code{NULL}, \code{"covid"} and \code{"poll"}. The default is \code{NULL} which indicates the
data is not specific to any supported use case.}

\item{family}{Character string specifying the distribution family for outcome measures.
Options are \code{"binomial"} and \code{"normal"}.}
}
\value{
A \code{data.frame} object.
}
\description{
Return example data based on the specified characteristics.
}
