% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{install_cmdstan}
\alias{install_cmdstan}
\title{Install CmdStan}
\usage{
install_cmdstan()
}
\description{
Installs CmdStan, the command-line interface for Stan, which is required for building statistical models in this package.
}
