% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{mrp_workflow}
\alias{mrp_workflow}
\title{Create a new MRPWorkflow object}
\usage{
mrp_workflow()
}
\value{
A \link{'MRPWorkflow'} object.
}
\description{
Create a new \code{\link{MRPWorkflow}} object that implements
the Bayesian data analysis workflow that underpins the application of
Multilevel Regression and Post-stratification (MRP).
}
