% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-check_fit_exists}
\alias{MRPModel-method-check_fit_exists}
\alias{check_fit_exists}
\title{Check if model has been fitted}
\usage{
check_fit_exists()
}
\description{
Checks whether the model has been fitted and results are available.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
