% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-diagnostics}
\alias{MRPModel-method-diagnostics}
\alias{diagnostics}
\title{Return sampling diagnostics}
\usage{
diagnostics(summarize = TRUE)
}
\arguments{
\item{summarize}{Logical indicating whether to return a summarized version of the diagnostics (default is TRUE)}
}
\description{
Retrieves MCMC diagnostics including convergence statistics and sampling efficiency measures.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_estimate_exists}},
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-formula}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-metadata}},
\code{\link{MRPModel-method-model_spec}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-stan_code}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
