% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-log_lik}
\alias{MRPModel-method-log_lik}
\alias{log_lik}
\title{Create inputs for leave-one-out cross-validation}
\usage{
log_lik()
}
\value{
A data.frame object containing log-likelihood values.
}
\description{
The \verb{$log_lik()} method of an \code{MRPModel} object
runs Stan's standalone generated quantities
and extracts log-likelihood values for leave-one-out cross-validation. This
method is called by the \verb{$compare_models()} method of an \code{MRPWorkflow} object
and does not need to be called directly by users.
}
