% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-loo}
\alias{MRPModel-method-loo}
\alias{loo}
\title{Create inputs for leave-one-out cross-validation}
\usage{
loo()
}
\description{
The \verb{$loo()} method runs Stan's standalone generated quantities
and extracts log-likelihood values for leave-one-out cross-validation.
}
