% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-metadata}
\alias{MRPModel-method-metadata}
\alias{metadata}
\title{Return model metadata}
\usage{
metadata()
}
\description{
The \verb{$metadata()} method of a \code{MRPModel} object
returns the metadata associated with the model,
including metadata inherited from a workflow object and model fitting parameters.
}
\examples{
if (FALSE) {
 library(shinymrp)

 # Load example data
 example_model <- example_model()

 # Get model metadata
 metadata <- example_model$metadata()
}

}
