% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-model_spec}
\alias{MRPModel-method-model_spec}
\alias{model_spec}
\title{Return model specification}
\usage{
model_spec()
}
\value{
A list containing the model specification including intercept, fixed effects, varying effects, and interactions.
}
\description{
The \verb{$model_spec()} method of a \code{MRPModel} object
returns the model specification list.
}
\examples{
if (FALSE) {
 library(shinymrp)

 # Load example data
 example_model <- example_model()

 # Get model specification list
 spec <- example_model$model_spec()
}

}
