% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-poststratify}
\alias{MRPModel-method-poststratify}
\alias{poststratify}
\title{Run poststratification to generate population estimates}
\usage{
poststratify(interval = 0.95)
}
\arguments{
\item{interval}{Confidence interval (a numeric value between 0 and 1) or
standard deviation (\code{"1sd"} or \code{"2sd"}) for the estimates (default is 0.95).}
}
\value{
A data.frame object containing the poststratified estimates and their
corresponding uncertainty intervals.
}
\description{
The \verb{$poststratify()} method of a \code{MRPModel} object
runs Stan's standalone generated quantities and extracts posterior samples
for poststratified estimates. This method is called by the \verb{$poststratify()}
method of a \code{MRPWorkflow} object and does not need to be called directly by users.
}
