% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-ppc}
\alias{MRPModel-method-ppc}
\alias{ppc}
\title{Run posterior predictive check}
\usage{
ppc()
}
\description{
The \verb{$ppc()} method runs Stan's standalone generated quantities
for posterior predictive checks.
}
