% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-save}
\alias{MRPModel-method-save}
\alias{save}
\title{Save model object to file}
\usage{
save(file)
}
\arguments{
\item{file}{File path where the model should be saved}
}
\description{
Saves a fitted MRPModel object to a file for later use.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
