% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-stan_code}
\alias{MRPModel-method-stan_code}
\alias{stan_code}
\title{Return model Stan code.}
\usage{
stan_code()
}
\value{
A character string containing the model Stan code.
}
\description{
The \verb{$stan_code()} method of an \code{MRPModel} object
returns the model Stan code. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#stan_code}{More examples of R6 classes}
vignette for usage examples.
}
