% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-summary}
\alias{MRPModel-method-summary}
\alias{summary}
\title{Return posterior summary table}
\usage{
summary()
}
\value{
A data frame containing the summary statistics for the posterior samples and diagnostics.
}
\description{
The \verb{$summary()} method returns a summary of posterior samples for the model parameters and diagnostics.
}
