% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-compare_models}
\alias{MRPWorkflow-method-compare_models}
\alias{compare_models}
\title{Compare models using LOO-CV}
\usage{
compare_models(..., suppress = NULL)
}
\arguments{
\item{...}{Multiple MRPModel objects to compare.}

\item{suppress}{Character string specifying output to suppress during comparison.}
}
\value{
A data.frame object containing the comparison results.
}
\description{
The \verb{$compare_models()} method compares multiple fitted \code{MRPModel} objects
using leave-one-out cross-validation to assess relative model performance. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#compare_models}{More examples of R6 classes}
vignette for usage examples.
}
