% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-estimate_map}
\alias{MRPWorkflow-method-estimate_map}
\alias{estimate_map}
\title{Create a choropleth map of MRP estimates}
\usage{
estimate_map(
  model,
  geo = NULL,
  time_index = NULL,
  interval = 0.95,
  file = NULL,
  ...
)
}
\arguments{
\item{model}{Fitted MRPModel object}

\item{geo}{Character string specifying the geographic level for mapping}

\item{time_index}{Numeric value specifying the time index for time-varying data}

\item{interval}{Confidence interval or standard deviation for the estimates (default is 0.95)}

\item{file}{Optional file path to save the map}

\item{...}{Additional arguments}
}
\value{
A highcharter map object showing MRP estimates by geography
}
\description{
Creates interactive choropleth maps showing MRP estimates by geographic regions.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-metadata}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
