% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-load_pstrat}
\alias{MRPWorkflow-method-load_pstrat}
\alias{load_pstrat}
\title{Load custom poststratification data}
\usage{
load_pstrat(pstrat_data, is_aggregated = TRUE)
}
\arguments{
\item{pstrat_data}{An object of class \code{data.frame} (or one that can be coerced to that class).}

\item{is_aggregated}{Logical indicating whether the poststratification data is already aggregated.}
}
\value{
No return value, called for side effects.
}
\description{
The \verb{$load_pstrat()} method processes and stores input poststratification data.
The object is subject to the same data preprocessing steps as the sample data. See the
\href{https://mrp-interface.github.io/shinymrp/articles/data-prep}{More on data preparation}
vignette for more information on data processing. For usage examples, refer to the
\href{https://mrp-interface.github.io/shinymrp/articles/example#load_pstrat}{More examples of R6 classes}
vignette.
}
