% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-outcome_plot}
\alias{MRPWorkflow-method-outcome_plot}
\alias{outcome_plot}
\title{Create summary plots of the outcome measure}
\usage{
outcome_plot(file = NULL, ...)
}
\arguments{
\item{file}{Optional file path to save the plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggsave}}, such as \code{width} and \code{height}.}
}
\value{
A ggplot object showing the outcome measure distribution.
}
\description{
The \verb{$outcome_plot()} method creates plots of the average outcome values.
Check out the \href{https://mrp-interface.github.io/shinymrp/articles/example#outcome_plot}{More examples of R6 classes}
vignette for usage examples.
}
