% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-pp_check}
\alias{MRPWorkflow-method-pp_check}
\alias{pp_check}
\title{Perform posterior predictive check}
\usage{
pp_check(model, file = NULL, ...)
}
\arguments{
\item{model}{Fitted MRPModel object.}

\item{file}{Optional file path to save the plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggsave}}, such as \code{width} and \code{height}.}
}
\value{
A ggplot object showing the posterior predictive check result.
}
\description{
The \verb{$pp_check()} method creates posterior predictive check plots
to assess model fit by comparing observed data to replicated data from
the posterior predictive distribution. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#pp_check}{More examples of R6 classes}
vignette for usage examples.
}
