% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-preprocess}
\alias{MRPWorkflow-method-preprocess}
\alias{preprocess}
\title{Preprocess sample data}
\usage{
preprocess(
  data,
  is_timevar = FALSE,
  is_aggregated = FALSE,
  special_case = NULL,
  family = NULL,
  time_freq = NULL,
  zip_threshold = 0,
  state_threshold = 0
)
}
\arguments{
\item{data}{Input sample data to be preprocessed}

\item{is_timevar}{Logical indicating whether the data contains time-varying components}

\item{is_aggregated}{Logical indicating whether the data is already aggregated}

\item{special_case}{Character string specifying special case handling (e.g., "covid", "poll")}

\item{family}{Character string specifying the model family (e.g., "binomial", "normal")}

\item{time_freq}{Character string specifying the time indexing frequency (e.g., "day", "month")}

\item{zip_threshold}{Numeric value specifying the minimum number of records required for a ZIP code to be included in the analysis (default is 0)}

\item{state_threshold}{Numeric value specifying the minimum number of records required for a state to be included in the analysis (default is 0)}
}
\description{
Preprocesses the input sample data by cleaning, validating, and preparing it for MRP analysis. This includes handling time-varying data, aggregated data, and special cases.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
