% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-sample_size_map}
\alias{MRPWorkflow-method-sample_size_map}
\alias{sample_size_map}
\title{Create sample size map}
\usage{
sample_size_map(file = NULL)
}
\arguments{
\item{file}{Optional file path to save the plot.}
}
\value{
A highcharter map object showing sample size distribution.
}
\description{
The \verb{$sample_size_map()} method creates interactive choropleth maps
showing data distribution with respect to geography. This method cannot be used
if there is no geographic information in either the sample or poststratification data.
}
