% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymrp-package.R
\docType{package}
\name{shinymrp-package}
\alias{shinymrp}
\alias{shinymrp-package}
\title{shinymrp: MRP Interface}
\description{
A graphical and a programmatic interface designed for intuitive applications of Multilevel Regression and Poststratification (MRP).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrp-interface.github.io/shinymrp/}
  \item Report bugs at \url{https://github.com/mrp-interface/shinymrp/issues}
}

}
\author{
\strong{Maintainer}: Toan Tran \email{trannttoan97@gmail.com}

Authors:
\itemize{
  \item Jonah Gabry \email{jgabry@gmail.com}
  \item Yajuan Si \email{yajuan@umich.edu}
}

}
\keyword{internal}
