% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_model}
\alias{example_model}
\title{Return example \code{MRPModel} object with estimation results}
\usage{
example_model()
}
\value{
A \code{MRPModel} object.
}
\description{
Return example \code{MRPModel} object with estimation results.
}
