% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_pstrat_data}
\alias{example_pstrat_data}
\title{Create example poststratification data
#' @description Generates example poststratification data for MRP analysis.}
\usage{
example_pstrat_data()
}
\value{
An example poststratification table.
}
\description{
Create example poststratification data
#' @description Generates example poststratification data for MRP analysis.
}
