% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_pstrat_data}
\alias{example_pstrat_data}
\title{Return example poststratification data}
\usage{
example_pstrat_data()
}
\value{
A \code{data.frame} object.
}
\description{
Return example poststratification data accepted by
the \verb{$load_pstrat()} method of a \code{MRPWorkflow} object.
}
