% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-check_estimate_exists}
\alias{MRPModel-method-check_estimate_exists}
\alias{check_estimate_exists}
\title{Check if poststratification has been performed}
\usage{
check_estimate_exists()
}
\value{
Logical indicating whether poststratification has been performed.
}
\description{
The \verb{$check_estimate_exists()} method checks whether poststratification has been performed.
}
