% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-check_fit_exists}
\alias{MRPModel-method-check_fit_exists}
\alias{check_fit_exists}
\title{Check if model has been fitted}
\usage{
check_fit_exists()
}
\value{
Logical indicating whether the model has been fitted.
}
\description{
The \verb{$check_fit_exists()} method of an \code{MRPModel} object
checks whether the model has been fitted. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#check_fit_exists}{More examples of R6 classes}
vignette for usage examples.
}
