% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-fit}
\alias{MRPModel-method-fit}
\alias{fit}
\title{Fit multilevel regression model using cmdstanr}
\usage{
fit(n_iter = 2000, n_chains = 4, seed = NULL, ...)
}
\arguments{
\item{n_iter}{Number of MCMC iterations per chain (including warmup iterations). Default is 2000.}

\item{n_chains}{Number of MCMC chains to run. Default is 4.}

\item{seed}{Random seed for reproducibility. Default is \code{NULL}.}

\item{...}{Additional arguments passed to CmdStanR \verb{$sample()} method}
}
\description{
The \verb{$fit()} method fits the model using Stan.
}
