% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-fit}
\alias{MRPModel-method-fit}
\alias{fit}
\title{Fit multilevel regression model using cmdstanr}
\usage{
fit(n_iter = 2000, n_chains = 4, seed = NULL, extra = NULL, ...)
}
\arguments{
\item{n_iter}{Number of MCMC iterations per chain}

\item{n_chains}{Number of MCMC chains to run}

\item{seed}{Random seed for reproducibility}

\item{extra}{Additional parameters for model fitting}

\item{...}{Additional arguments passed to CmdStanR's \code{sample()} method}
}
\description{
Fits the MRP model using Stan for Bayesian estimation with MCMC sampling.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-check_if_poststratified}},
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
