% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-formula}
\alias{MRPModel-method-formula}
\alias{formula}
\title{Return model formula}
\usage{
formula()
}
\value{
A character string of the model formula.
}
\description{
The \verb{$formula()} method of an \code{MRPModel} object
returns the lme4-style formula constructed from the given model specification.
Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#formula}{More examples of R6 classes}
vignette for usage examples.
}
