% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-formula}
\alias{MRPModel-method-formula}
\alias{formula}
\title{Return model formula}
\usage{
formula()
}
\description{
The \verb{$formula()} method of a \code{MRPModel} object
returns the lme4-style formula constructed from the model specification list.
}
\examples{
\donttest{
 library(shinymrp)

 # Load example data
 example_model <- example_model()

 # Get model formula
 formula <- example_model$formula()
}

}
