% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-loo}
\alias{MRPModel-method-loo}
\alias{loo}
\title{Create inputs for leave-one-out cross-validation}
\usage{
loo()
}
\description{
Runs leave-one-out cross-validation to assess model predictive performance and enable model comparison.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-check_if_poststratified}},
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
