% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-metadata}
\alias{MRPModel-method-metadata}
\alias{metadata}
\title{Return model metadata}
\usage{
metadata()
}
\description{
Retrieves the metadata associated with the model, including information about family, time variables, and fitting parameters.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_estimate_exists}},
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-formula}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-model_spec}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-stan_code}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
