% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-metadata}
\alias{MRPModel-method-metadata}
\alias{metadata}
\title{Return model metadata.}
\usage{
metadata()
}
\value{
A list containing the model metadata.
}
\description{
The \verb{$metadata()} method of an \code{MRPModel} object
returns the metadata associated with the model,
including metadata inherited from a workflow object and model fitting parameters.
Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#metadata}{More examples of R6 classes}
vignette for usage examples.
}
