% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-poststratify}
\alias{MRPModel-method-poststratify}
\alias{poststratify}
\title{Run post-stratification to generate population estimates}
\usage{
poststratify(interval = 0.95)
}
\arguments{
\item{interval}{Confidence interval or standard deviation for the estimates (default is 0.95)}
}
\description{
Runs post-stratification using the fitted model to generate population-level estimates across different subgroups and geographies.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-check_if_poststratified}},
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
