% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-poststratify}
\alias{MRPModel-method-poststratify}
\alias{poststratify}
\title{Run poststratification to generate population estimates}
\usage{
poststratify(interval = 0.95)
}
\arguments{
\item{interval}{Confidence interval (a numeric value between 0 and 1) or
standard deviation (\code{"1sd"}, \code{"2sd"}, or \code{"3sd"}) for the estimates (default is 0.95).}
}
\value{
A data.frame object containing the poststratified estimates and their
corresponding uncertainty intervals.
}
\description{
The \verb{$poststratify()} method runs Stan's standalone generated quantities
and extracts posterior samples for poststratified estimates.
}
