% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-ppc}
\alias{MRPModel-method-ppc}
\alias{ppc}
\title{Run posterior predictive check}
\usage{
ppc()
}
\description{
The \verb{$ppc()} method of a \code{MRPModel} object
runs Stan's standalone generated quantities
to draw from the posterior predictive distribution. This method is called
by the \verb{$pp_check()} method of a \code{MRPWorkflow} object and does
not need to be called directly by users.
}
