% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-ppc}
\alias{MRPModel-method-ppc}
\alias{ppc}
\title{Run posterior predictive check}
\usage{
ppc()
}
\description{
Runs posterior predictive checks to assess model fit by generating replicated data from the posterior predictive distribution.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-check_if_poststratified}},
\code{\link{MRPModel-method-code}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-summary}}
}
\concept{MRPModel methods}
