% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-save}
\alias{MRPModel-method-save}
\alias{save}
\title{Save model object to file}
\usage{
save(file)
}
\arguments{
\item{file}{File path where the model should be saved.}
}
\description{
The \verb{$save()} method of a \code{MRPModel} object
saves a fitted MRPModel object to a file for later use.
}
