% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-save}
\alias{MRPModel-method-save}
\alias{save}
\title{Save model object to file}
\usage{
save(file)
}
\arguments{
\item{file}{File path where the model should be saved.}
}
\value{
No return value, called for side effects.
}
\description{
The \verb{$save()} method of an \code{MRPModel} object
saves a fitted MRPModel object to a file for later use.
\code{qs::qsave()} is used internally, and it is customary
to use the \code{.qs} file extension. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#save}{More examples of R6 classes}
vignette for usage examples.
}
