% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-save}
\alias{MRPModel-method-save}
\alias{save}
\title{Save model object to file}
\usage{
save(file)
}
\arguments{
\item{file}{File path where the model should be saved.}
}
\value{
No return value, called for side effects.
}
\description{
The \verb{$save()} method of an \code{MRPModel} object
saves a fitted MRPModel object to a file for later use.
\code{qs::qsave()} is used internally, and it is customary
to use the \code{.qs} file extension.
}
\examples{
if (FALSE) {
 library(shinymrp)

 # Initialize workflow
 workflow <- mrp_workflow()

 # Load example model
 example_model <- example_model()

 # Save model to file
 # example_model$save("model.qs")

 # Load model later
 # loaded_model <- qs::qread("model.qs")
}

}
