% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-stan_code}
\alias{MRPModel-method-stan_code}
\alias{stan_code}
\title{Return model Stan code.}
\usage{
stan_code()
}
\value{
A character string containing the model Stan code.
}
\description{
The \verb{$stan_code()} method of an \code{MRPModel} object
returns the model Stan code.
}
\examples{
if (FALSE) {
 library(shinymrp)

 # Load example data
 example_model <- example_model()

 # Get Stan code
 stan_code <- example_model$stan_code()
}

}
