% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-summary}
\alias{MRPModel-method-summary}
\alias{summary}
\title{Return posterior summary table}
\usage{
summary()
}
\description{
Retrieves a summary of the fitted model parameters including posterior means, credible intervals, and diagnostics.
}
\seealso{
Other MRPModel methods: 
\code{\link{MRPModel-method-check_estimate_exists}},
\code{\link{MRPModel-method-check_fit_exists}},
\code{\link{MRPModel-method-diagnostics}},
\code{\link{MRPModel-method-fit}},
\code{\link{MRPModel-method-formula}},
\code{\link{MRPModel-method-loo}},
\code{\link{MRPModel-method-metadata}},
\code{\link{MRPModel-method-model_spec}},
\code{\link{MRPModel-method-poststratify}},
\code{\link{MRPModel-method-ppc}},
\code{\link{MRPModel-method-save}},
\code{\link{MRPModel-method-stan_code}}
}
\concept{MRPModel methods}
