% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-summary}
\alias{MRPModel-method-summary}
\alias{summary}
\title{Return posterior summary table}
\usage{
summary()
}
\value{
A list of data.frame objects containing posterior sample
summary and diagnostics for model parameters:
\itemize{
\item fixed effects (\code{fixed})
\item standard deviations of varying effects (\code{varying})
\item standard deviations of residuals (\code{residual})
\item BYM2 proportion of spatial structure (\code{spatial_proportion})
}
}
\description{
The \verb{$summary()} method of an \code{MRPModel} object
returns tables containing the summary of posterior samples
for the model parameters and diagnostics. Check out the
\href{https://mrp-interface.github.io/shinymrp/articles/example#summary}{More examples of R6 classes}
vignette for usage examples.
}
