% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{MRPModel-method-summary}
\alias{MRPModel-method-summary}
\alias{summary}
\title{Return posterior summary table}
\usage{
summary()
}
\value{
A list of \code{data.frame} objects containing posterior sample
summary and diagnostics for model parameters:
\itemize{
\item fixed effects (\code{fixed})
\item standard deviations of varying effects (\code{varying})
\item standard deviations of residuals (\code{other})
}
}
\description{
The \verb{$summary()} method of a \code{MRPModel} object
returns tables containing summary of posterior samples
for the model parameters and diagnostics.
}
