% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{MRPModel}
\alias{MRPModel}
\title{MRPModel objects}
\format{
An \link[R6:R6Class]{R6} generator object.
}
\description{
A \code{MRPModel} object is an \link[R6:R6Class]{R6} object created by the
\verb{$create_model()} method of a
\code{\link{MRPWorkflow}} object. Each \code{MRPModel} object represents a
multilevel regression model with methods for sampling, diagnostics,
and poststratification.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MRPModel-new}{\code{MRPModel$new()}}
\item \href{#method-MRPModel-effects}{\code{MRPModel$effects()}}
\item \href{#method-MRPModel-formula}{\code{MRPModel$formula()}}
\item \href{#method-MRPModel-mrp}{\code{MRPModel$mrp()}}
\item \href{#method-MRPModel-metadata}{\code{MRPModel$metadata()}}
\item \href{#method-MRPModel-plotdata}{\code{MRPModel$plotdata()}}
\item \href{#method-MRPModel-linkdata}{\code{MRPModel$linkdata()}}
\item \href{#method-MRPModel-stan_data}{\code{MRPModel$stan_data()}}
\item \href{#method-MRPModel-stan_code}{\code{MRPModel$stan_code()}}
\item \href{#method-MRPModel-fit}{\code{MRPModel$fit()}}
\item \href{#method-MRPModel-check_fit_exists}{\code{MRPModel$check_fit_exists()}}
\item \href{#method-MRPModel-code}{\code{MRPModel$code()}}
\item \href{#method-MRPModel-summary}{\code{MRPModel$summary()}}
\item \href{#method-MRPModel-diagnostics}{\code{MRPModel$diagnostics()}}
\item \href{#method-MRPModel-ppc}{\code{MRPModel$ppc()}}
\item \href{#method-MRPModel-loo}{\code{MRPModel$loo()}}
\item \href{#method-MRPModel-poststratify}{\code{MRPModel$poststratify()}}
\item \href{#method-MRPModel-save}{\code{MRPModel$save()}}
\item \href{#method-MRPModel-clone}{\code{MRPModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the MRPModel class with specified effects, data, and metadata for Bayesian model fitting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$new(effects, mrp, metadata, linkdata, plotdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effects}}{List containing model effects specification including intercept, fixed effects, varying effects, and interactions}

\item{\code{mrp}}{List containing the MRP data structure with input sample data and new poststratification data}

\item{\code{metadata}}{List containing metadata about the analysis including family, time variables, and special cases}

\item{\code{linkdata}}{List containing information about data linking including geography and ACS year}

\item{\code{plotdata}}{List containing data prepared for visualization including dates and geojson objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{MRPModel} object initialized with the provided effects, MRP data, metadata, link data, and plot data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-effects"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-effects}{}}}
\subsection{Method \code{effects()}}{
Retrieves the effects specification used in the model, including intercept, fixed effects, varying effects, and interactions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$effects()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-formula"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-formula}{}}}
\subsection{Method \code{formula()}}{
Retrieves the model formula constructed from the effects specification.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$formula()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-mrp"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-mrp}{}}}
\subsection{Method \code{mrp()}}{
Retrieves the MRP data structure containing input sample data and poststratification data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$mrp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-metadata}{}}}
\subsection{Method \code{metadata()}}{
Retrieves the metadata associated with the model, including information about family, time variables, and fitting parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-plotdata"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-plotdata}{}}}
\subsection{Method \code{plotdata()}}{
Retrieves the data prepared for visualization, including dates and geojson objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$plotdata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-linkdata"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-linkdata}{}}}
\subsection{Method \code{linkdata()}}{
Retrieves the data linking information including geography and ACS year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$linkdata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-stan_data"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-stan_data}{}}}
\subsection{Method \code{stan_data()}}{
Retrieves the Stan data structure used for MCMC sampling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$stan_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-stan_code"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-stan_code}{}}}
\subsection{Method \code{stan_code()}}{
Retrieves Stan code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$stan_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the MRP model using Stan for Bayesian estimation with MCMC sampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$fit(n_iter = 2000, n_chains = 4, seed = NULL, extra = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_iter}}{Number of MCMC iterations per chain}

\item{\code{n_chains}}{Number of MCMC chains to run}

\item{\code{seed}}{Random seed for reproducibility}

\item{\code{extra}}{Additional parameters for model fitting}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-check_fit_exists"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-check_fit_exists}{}}}
\subsection{Method \code{check_fit_exists()}}{
Checks whether the model has been fitted and results are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$check_fit_exists()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-code"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-code}{}}}
\subsection{Method \code{code()}}{
Retrieves the Stan model code used for MCMC fitting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-summary"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-summary}{}}}
\subsection{Method \code{summary()}}{
Retrieves a summary of the fitted model parameters including posterior means, credible intervals, and diagnostics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-diagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-diagnostics}{}}}
\subsection{Method \code{diagnostics()}}{
Retrieves MCMC diagnostics including convergence statistics and sampling efficiency measures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$diagnostics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-ppc"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-ppc}{}}}
\subsection{Method \code{ppc()}}{
Runs posterior predictive checks to assess model fit by generating replicated data from the posterior predictive distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$ppc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-loo"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-loo}{}}}
\subsection{Method \code{loo()}}{
Runs leave-one-out cross-validation to assess model predictive performance and enable model comparison.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$loo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-poststratify"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-poststratify}{}}}
\subsection{Method \code{poststratify()}}{
Runs poststratification using the fitted model to generate population-level estimates across different subgroups and geographies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$poststratify(interval = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interval}}{Confidence interval or standard deviation for the estimates (default is 0.95)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-save"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-save}{}}}
\subsection{Method \code{save()}}{
Saves a fitted MRPModel object to a file for later use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$save(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{File path where the model should be saved}

\item{\code{model}}{Fitted MRPModel object to save}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MRPModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
