% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-covar_hist}
\alias{MRPWorkflow-method-covar_hist}
\alias{covar_hist}
\title{Create geographic covariate distribution histogram}
\usage{
covar_hist(covar, file = NULL, ...)
}
\arguments{
\item{covar}{Character string specifying the covariate. Options: "edu",
"poverty", "employ", "income", "urban", "adi"}

\item{file}{Optional file path to save the plot}

\item{...}{Additional arguments passed to ggsave}
}
\value{
A ggplot object showing the covariate distribution histogram
}
\description{
Creates histogram plots showing the distribution of geographic
covariates across zip codes. Only available for COVID data.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
