% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-covar_hist}
\alias{MRPWorkflow-method-covar_hist}
\alias{covar_hist}
\title{Create geographic covariate distribution histogram}
\usage{
covar_hist(covar, file = NULL, ...)
}
\arguments{
\item{covar}{Character string specifying the geographic covariate. Options are
\code{"college"}, \code{"poverty"}, \code{"employment"}, \code{"income"}, \code{"urbanicity"}, and \code{"adi"}.}

\item{file}{Optional file path to save the plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggsave}}, such as \code{width} and \code{height}.}
}
\value{
A ggplot object showing the covariate distribution histogram.
}
\description{
The \code{covar_hist()} method creates histogram plots
showing the distribution of geographic covariates across ZIP codes. Refer to the
\href{https://mrp-interface.github.io/shinymrp/articles/data-prep#geographic-identifiers-and-covariates}{More on data preparation} for their definitions.
This method is only available for \href{https://mrp-interface.github.io/shinymrp/articles/data-prep#data-modules}{COVID data}.
}
\examples{
if (FALSE) {
library(shinymrp)
 library(shinymrp)

 # Initialize workflow
 workflow <- mrp_workflow()

 # Load example data
 sample_data <- example_sample_data()

 # Preprocess sample data
 workflow$preprocess(
   sample_data,
   is_timevar = TRUE,
   is_aggregated = TRUE,
   special_case = "covid",
   family = "binomial"
 )

 # Link to ACS data at ZIP code level
 workflow$link_acs(
   link_geo = "zip",
   acs_year = 2021
 )

 # Create and (optionally) save a covariate distribution histogram
 # with dimensions in inches
 college_hist <- workflow$covar_hist(
   "college",
   file = "path/to/college_hist.png",
   width = 18,
   height = 8,
   units = "in"
 )
}

}
