% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-create_model}
\alias{MRPWorkflow-method-create_model}
\alias{create_model}
\title{Create a MRPModel object}
\usage{
create_model(model_spec)
}
\arguments{
\item{model_spec}{List containing model effects specification including intercept, fixed effects, varying effects, and interactions}
}
\value{
A new MRPModel object
}
\description{
The \verb{$create_model()} method creates a new MRPModel object
with Stan code generated from the model specification list.
\code{CmdStanR} objects are used internally for model fitting.
}
