% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-estimate_plot}
\alias{MRPWorkflow-method-estimate_plot}
\alias{estimate_plot}
\title{Visualize estimates for demographic groups}
\usage{
estimate_plot(
  model,
  group = NULL,
  interval = 0.95,
  show_caption = TRUE,
  file = NULL,
  ...
)
}
\arguments{
\item{model}{Fitted MRPModel object}

\item{group}{Character string specifying the demographic group for plotting}

\item{interval}{Confidence interval (a numeric value between 0 and 1) or
standard deviation (\code{"1sd"}, \code{"2sd"}, or \code{"3sd"}) for the estimates (default is 0.95).}

\item{show_caption}{Logical indicating whether to show the caption in the plot (default is TRUE).}

\item{file}{Optional file path to save the plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggsave}}.}
}
\value{
A ggplot object showing MRP estimates.
}
\description{
The \verb{$estimate_plot()} method creates plots showing overall MRP estimates or
estimates for different demographic groups.
}
\examples{
\donttest{
 library(shinymrp)

 # Initialize workflow
 workflow <- mrp_workflow()

 # Load example data
 example_model <- example_model()

 # Overall estimates
 workflow$estimate_plot(example_model)

 # Estimates by demographic group
 workflow$estimate_plot(example_model, group = "sex")
}

}
