% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-estimate_plot}
\alias{MRPWorkflow-method-estimate_plot}
\alias{estimate_plot}
\title{Visualize estimates for demographic groups}
\usage{
estimate_plot(
  model,
  group = NULL,
  interval = 0.95,
  show_caption = TRUE,
  file = NULL,
  ...
)
}
\arguments{
\item{model}{Fitted MRPModel object}

\item{group}{Character string specifying the demographic group for plotting}

\item{interval}{Confidence interval or standard deviation for the estimates (default is 0.95)}

\item{show_caption}{Logical indicating whether to show the caption in the plot (default is TRUE)}

\item{file}{Optional file path to save the plot}

\item{...}{Additional arguments passed to ggsave}
}
\value{
A ggplot object showing the group estimates
}
\description{
Creates plots showing MRP estimates for different subgroups, either over time (for time-varying data) or as static estimates (for cross-sectional data).
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
