% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-link_acs}
\alias{MRPWorkflow-method-link_acs}
\alias{link_acs}
\title{Link sample data to ACS data}
\usage{
link_acs(link_geo = NULL, acs_year = 2023)
}
\arguments{
\item{link_geo}{Character string specifying the geographic level for linking (e.g., "state", "county", "zip")}

\item{acs_year}{Numeric value specifying the ACS year to use for poststratification data}
}
\description{
Links the preprocessed sample data to ACS poststratification data based on geographic and demographic variables.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
