% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-load_pstrat}
\alias{MRPWorkflow-method-load_pstrat}
\alias{load_pstrat}
\title{Load custom post-stratification data}
\usage{
load_pstrat(pstrat_data, is_aggregated = FALSE)
}
\arguments{
\item{pstrat_data}{Custom poststratification data to be loaded}

\item{is_aggregated}{Logical indicating whether the poststratification data is already aggregated}
}
\description{
Loads and processes custom poststratification data instead of using ACS data. This method validates the data and prepares it for MRP analysis.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-metadata}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
