% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-load_pstrat}
\alias{MRPWorkflow-method-load_pstrat}
\alias{load_pstrat}
\title{Load custom poststratification data}
\usage{
load_pstrat(pstrat_data, is_aggregated = FALSE)
}
\arguments{
\item{pstrat_data}{An object of class \code{data.frame} (or one that can be coerced to that class).}

\item{is_aggregated}{Logical indicating whether the poststratification data is already aggregated.}
}
\description{
The \verb{$load_pstrat()} method processes and stores input poststratification data.
The object is subject to the same data preprocessing steps as the sample data.
}
\examples{
\dontrun{
 library(shinymrp)

 # Initialize workflow
 workflow <- mrp_workflow()

 # Load example sample data
 sample_data <- example_sample_data()

 # Preprocess sample data
 workflow$preprocess(
   sample_data,
   is_timevar = TRUE,
   is_aggregated = TRUE,
   special_case = NULL,
   family = "binomial"
 )

 # Load example poststratification data
 pstrat_data <- example_pstrat_data()

 # Read in poststratification data
 workflow$load_pstrat(
   pstrat_data = pstrat_data,
   is_aggregated = TRUE
 )
}
}
