% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-load_pstrat}
\alias{MRPWorkflow-method-load_pstrat}
\alias{load_pstrat}
\title{Load custom poststratification data}
\usage{
load_pstrat(pstrat_data, is_aggregated = FALSE)
}
\arguments{
\item{pstrat_data}{An object of class \code{data.frame} (or one that can be coerced to that class).}

\item{is_aggregated}{Logical indicating whether the poststratification data is already aggregated.}
}
\description{
The \verb{$load_pstrat()} method processes and stores input poststratification data.
The object is subject to the same data preprocessing steps as the sample data.
}
