% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-outcome_map}
\alias{MRPWorkflow-method-outcome_map}
\alias{outcome_map}
\title{Visualize raw outcome measure by geography}
\usage{
outcome_map(summary_type = "max", file = NULL)
}
\arguments{
\item{summary_type}{Character string for time-varying data, indicating whether to display the
highest (\code{"max"}) or lowest (\code{"min"}) temporal average. Default is \code{"max"}.}

\item{file}{Optional file path to save the map.}
}
\value{
A highcharter map object showing outcome measures by geography.
}
\description{
The \verb{$outcome_map()} method creates maps showing average outcome measure by geography for
cross-sectional data, or highest/lowest temporal average for time-varying data. The sample and
poststratification data must contain geographic information for this method to work.
}
