% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-outcome_plot}
\alias{MRPWorkflow-method-outcome_plot}
\alias{outcome_plot}
\title{Create summary plots of the outcome measure}
\usage{
outcome_plot(file = NULL, ...)
}
\arguments{
\item{file}{Optional file path to save the plot}

\item{...}{Additional arguments passed to ggsave}
}
\value{
A ggplot object showing the outcome measure distribution
}
\description{
Creates plots showing the distribution of outcome measures over time (for time-varying data) or as static distributions (for cross-sectional data).
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-metadata}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
