% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-pp_check}
\alias{MRPWorkflow-method-pp_check}
\alias{pp_check}
\title{Perform posterior predictive check}
\usage{
pp_check(model, file = NULL, ...)
}
\arguments{
\item{model}{Fitted MRPModel object.}

\item{file}{Optional file path to save the plot.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggsave}}, such as \code{width} and \code{height}.}
}
\description{
The \verb{$pp_check()} method creates posterior predictive check plots
to assess model fit by comparing observed data to replicated data from
the posterior predictive distribution.
}
\examples{
if (FALSE) {
 library(shinymrp)

 # Initialize workflow
 workflow <- mrp_workflow()

 # Load example data
 example_model <- example_model()

 # Perform posterior predictive check and (optionally) save the plot
 ppc_plot <- workflow$pp_check(example_model, file = "/path/to/ppc_plot.png")
}

}
