% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-preprocess}
\alias{MRPWorkflow-method-preprocess}
\alias{preprocess}
\title{Preprocess sample data}
\usage{
preprocess(
  data,
  is_timevar = FALSE,
  is_aggregated = FALSE,
  special_case = NULL,
  family = NULL,
  time_freq = NULL,
  freq_threshold = 0
)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced to that class).}

\item{is_timevar}{Logical indicating whether the data contains time-varying components.}

\item{is_aggregated}{Logical indicating whether the data is already aggregated.
See the \href{https://mrp-interface.github.io/shinymrp/articles/data-prep}{"More on data preparation"} vignette for more information.}

\item{special_case}{Character string specifying special case handling. Options are \code{NULL} (the default), \code{"covid"}, and \code{"poll"}.
See the \href{https://mrp-interface.github.io/shinymrp/articles/data-prep}{"More on data preparation"} vignette for more information.}

\item{family}{Character string specifying the distribution family for the outcome variable. Options are \code{"binomial"} for binary outcome measures and \code{"normal"} for continuous outcome measures.}

\item{time_freq}{Character string specifying the time indexing frequency or time length for grouping dates (YYYY-MM-DD) in the data.
Options are \code{NULL} (the default), \code{"week"}, \code{"month"}, and \code{"year"}. This parameter must be \code{NULL} for cross-sectional data
or time-varying data that already has time indices.}

\item{freq_threshold}{Numeric value specifying the minimum frequency threshold for including observations.
Values with lower frequency will cause the entire row to be removed. The default value is 0.}
}
\description{
The \verb{$preprocess()} method runs the preprocessing pipeline
that includes data standardization, filtering, imputation, and aggregation.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocessed_data}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
