% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-preprocessed_data}
\alias{MRPWorkflow-method-preprocessed_data}
\alias{preprocessed_data}
\title{Return preprocessed sample data}
\usage{
preprocessed_data()
}
\value{
A data.frame object containing the preprocessed sample data.
}
\description{
The \verb{$preprocessed_data()} method returns the preprocessed sample data.
Check out the \href{https://mrp-interface.github.io/shinymrp/articles/example#preprocessed_data}{More examples of R6 classes}
vignette for usage examples.
}
