% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-preprocessed_data}
\alias{MRPWorkflow-method-preprocessed_data}
\alias{preprocessed_data}
\title{Return preprocessed sample data}
\usage{
preprocessed_data()
}
\description{
Retrieves the preprocessed sample data that has been prepared for MRP analysis.
}
\seealso{
Other MRPWorkflow methods: 
\code{\link{MRPWorkflow-method-compare_models}},
\code{\link{MRPWorkflow-method-covar_hist}},
\code{\link{MRPWorkflow-method-create_model}},
\code{\link{MRPWorkflow-method-demo_bars}},
\code{\link{MRPWorkflow-method-estimate_map}},
\code{\link{MRPWorkflow-method-estimate_plot}},
\code{\link{MRPWorkflow-method-link_acs}},
\code{\link{MRPWorkflow-method-load_pstrat}},
\code{\link{MRPWorkflow-method-outcome_map}},
\code{\link{MRPWorkflow-method-outcome_plot}},
\code{\link{MRPWorkflow-method-pp_check}},
\code{\link{MRPWorkflow-method-preprocess}},
\code{\link{MRPWorkflow-method-sample_size_map}}
}
\concept{MRPWorkflow methods}
