% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow-method-sample_size_map}
\alias{MRPWorkflow-method-sample_size_map}
\alias{sample_size_map}
\title{Create sample size map}
\usage{
sample_size_map(file = NULL)
}
\arguments{
\item{file}{Optional file path with .html extension to save the interactive map.
Expand the hamburger menu in the top right corner of the map to access other export options.}
}
\value{
A highcharter map object showing sample size distribution.
}
\description{
The \verb{$sample_size_map()} method creates interactive choropleth maps
showing data distribution with respect to geography. This method cannot be used
if either the sample or the poststratification data contains no geographic information.
Check out the \href{https://mrp-interface.github.io/shinymrp/articles/example#sample_size_map}{More examples of R6 classes}
vignette for usage examples.
}
