% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{MRPWorkflow}
\alias{MRPWorkflow}
\title{Create demographic comparison bar plots}
\description{
Create demographic comparison bar plots

Create demographic comparison bar plots
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MRPWorkflow-new}{\code{MRPWorkflow$new()}}
\item \href{#method-MRPWorkflow-metadata}{\code{MRPWorkflow$metadata()}}
\item \href{#method-MRPWorkflow-data}{\code{MRPWorkflow$data()}}
\item \href{#method-MRPWorkflow-preprocess}{\code{MRPWorkflow$preprocess()}}
\item \href{#method-MRPWorkflow-link_acs}{\code{MRPWorkflow$link_acs()}}
\item \href{#method-MRPWorkflow-load_pstrat}{\code{MRPWorkflow$load_pstrat()}}
\item \href{#method-MRPWorkflow-demo_bars}{\code{MRPWorkflow$demo_bars()}}
\item \href{#method-MRPWorkflow-covar_hist}{\code{MRPWorkflow$covar_hist()}}
\item \href{#method-MRPWorkflow-sample_size_map}{\code{MRPWorkflow$sample_size_map()}}
\item \href{#method-MRPWorkflow-outcome_plot}{\code{MRPWorkflow$outcome_plot()}}
\item \href{#method-MRPWorkflow-outcome_map}{\code{MRPWorkflow$outcome_map()}}
\item \href{#method-MRPWorkflow-check_effects}{\code{MRPWorkflow$check_effects()}}
\item \href{#method-MRPWorkflow-create_model}{\code{MRPWorkflow$create_model()}}
\item \href{#method-MRPWorkflow-load_model}{\code{MRPWorkflow$load_model()}}
\item \href{#method-MRPWorkflow-check_model}{\code{MRPWorkflow$check_model()}}
\item \href{#method-MRPWorkflow-compare_models}{\code{MRPWorkflow$compare_models()}}
\item \href{#method-MRPWorkflow-pp_check}{\code{MRPWorkflow$pp_check()}}
\item \href{#method-MRPWorkflow-estimate_plot}{\code{MRPWorkflow$estimate_plot()}}
\item \href{#method-MRPWorkflow-estimate_map}{\code{MRPWorkflow$estimate_map()}}
\item \href{#method-MRPWorkflow-clone}{\code{MRPWorkflow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-new"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-metadata}{}}}
\subsection{Method \code{metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-data"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-data}{}}}
\subsection{Method \code{data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$preprocess(
  data,
  is_timevar = FALSE,
  is_aggregated = FALSE,
  special_case = NULL,
  family = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-link_acs"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-link_acs}{}}}
\subsection{Method \code{link_acs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$link_acs(link_geo = NULL, acs_year = 2023)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-load_pstrat"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-load_pstrat}{}}}
\subsection{Method \code{load_pstrat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$load_pstrat(pstrat_data, is_aggregated = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-demo_bars"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-demo_bars}{}}}
\subsection{Method \code{demo_bars()}}{
Creates bar plots comparing demographic distributions between
input survey data and target population data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$demo_bars(demo, file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{separate}}{Logical indicating whether to create separate plots (TRUE) or
side-by-side comparison (FALSE)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object or patchwork object showing demographic comparisons
Create geographic covariate distribution histogram
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-covar_hist"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-covar_hist}{}}}
\subsection{Method \code{covar_hist()}}{
Creates histogram plots showing the distribution of geographic
covariates across zip codes. Only available for COVID data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$covar_hist(covar, file = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariate}}{Character string specifying the covariate. Options: "edu",
"poverty", "employ", "income", "urban", "adi"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object showing the covariate distribution histogram
Create sample size map
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-sample_size_map"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-sample_size_map}{}}}
\subsection{Method \code{sample_size_map()}}{
Creates interactive choropleth maps showing data distribution
with respect to geography.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$sample_size_map(file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{geo}}{Character string specifying geographic level: "county" or "state"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A highcharter map object showing sample size distribution
Create outcome measure maps by geography
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-outcome_plot"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-outcome_plot}{}}}
\subsection{Method \code{outcome_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$outcome_plot(file = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-outcome_map"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-outcome_map}{}}}
\subsection{Method \code{outcome_map()}}{
Creates maps showing average outcome measure by geography for
cross-sectional data, or highest/lowest weekly average for time-varying data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$outcome_map(summary_type = NULL, file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{summary_type}}{Character string for time-varying data: "max" or "min"}

\item{\code{geo}}{Character string specifying geographic level: "county" or "state"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A highcharter map object showing outcome measures by geography
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-check_effects"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-check_effects}{}}}
\subsection{Method \code{check_effects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$check_effects(effects)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-create_model"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-create_model}{}}}
\subsection{Method \code{create_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$create_model(effects)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-load_model"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-load_model}{}}}
\subsection{Method \code{load_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$load_model(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-check_model"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-check_model}{}}}
\subsection{Method \code{check_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$check_model(model)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-compare_models"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-compare_models}{}}}
\subsection{Method \code{compare_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$compare_models(..., suppress = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-pp_check"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-pp_check}{}}}
\subsection{Method \code{pp_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$pp_check(model, file = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-estimate_plot"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-estimate_plot}{}}}
\subsection{Method \code{estimate_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$estimate_plot(model, subgroup, file = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-estimate_map"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-estimate_map}{}}}
\subsection{Method \code{estimate_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$estimate_map(
  model,
  geo = NULL,
  time_index = NULL,
  file = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MRPWorkflow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MRPWorkflow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MRPWorkflow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
