% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymrp-package.R
\docType{package}
\name{shinymrp-package}
\alias{shinymrp}
\alias{shinymrp-package}
\title{shinymrp: Interface for MRP}
\description{
Dual interfaces, graphical and programmatic, designed for intuitive applications of Multilevel Regression and Poststratification (MRP). Users can apply the method to a variety of datasets, from electronic health records to sample survey data, through an end-to-end Bayesian data analysis workflow. The package provides robust tools for data cleaning, exploratory analysis, flexible model building, and insightful result visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrp-interface.github.io/shinymrp/}
  \item Report bugs at \url{https://github.com/mrp-interface/shinymrp/issues}
}

}
\author{
\strong{Maintainer}: Toan Tran \email{trannttoan97@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Jonah Gabry \email{jgabry@gmail.com} [copyright holder]
  \item Yajuan Si \email{yajuan@umich.edu} [copyright holder]
}

}
\keyword{internal}
